pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "A2A agent for validating natural language task planning backed by PDDL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "openai>=2.15.0",
    "pddlval",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
    "unified-planning>=1.3.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[tool.uv.sources]
pddlval = { git = "https://github.com/ElliotGestrin/pddlval.git" }
